캐시 관련 설정 분리
